preferred_format function moved to config.py
